add custom headers support for connection
